define repository and crawler

+++ fin-crawling-fast-api/server/app/crawler/MarcapCrawler.py
@@ -3,7 +3,7 @@ import asyncio
 import os
 
 from datetime import datetime, timedelta
-from typing import TypeVar
+from typing import TypeVar, Final
 
 from pymitter import EventEmitter
 from selenium import webdriver
@@ -12,29 +12,40 @@ from selenium.webdriver.remote.webdriver import WebDriver
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 from watchdog.events import FileCreatedEvent
+import uuid
 
-from fin_crawling.data.StockCrawlingDownloadTaskDTO import \
-    StockCrawlingDownloadTaskDTO
-from fin_crawling.data.StockCrawlingRunCrawlingDTO import \
-    StockCrawlingRunCrawlingDTO
-from fin_crawling.utils.FileUtil import changeCharSet
-from fin_crawling.data.StockMarketCapitalResultDTO import StockMarketCapitalResultDTO
+from app.observer.DownloadObserver import DownloadObserver
+from app.observer.CmdFileSystemEventHandler import FILE_SYSTEM_HANDLER
 
 from .Crawler import Crawler
+from pathlib import Path
+from app.model.dto import StockCrawlingDownloadTaskDTO, StockCrawlingRunCrawlingDTO, StockCrawlingTaskDTO, StockMarketCapitalResultDTO, StockMarketCapitalDTO
 
 T = TypeVar("T")
+
+EVENT_MARCAP_CRAWLING_ON_CONNECTING_WEBDRIVER: Final = "marcapCrawler/onConnectingWebdriver"
+EVENT_MARCAP_CRAWLING_ON_START_CRAWLING: Final = "marcapCrawler/onStartCrawling"
+EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_START: Final = "marcapCrawler/onDownloadStart"
+EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_COMPLETE: Final = "marcapCrawler/onDownloadComplete"
+EVENT_MARCAP_CRAWLING_ON_PARSING_COMPLETE: Final = "marcapCrawler/onParsingComplete"
 
 
 class MarcapCrawler(Crawler[T]):
     @staticmethod
     def create(ee: EventEmitter) -> MarcapCrawler:
         crawler = MarcapCrawler[None]()
-        crawler.ee = ee
+        crawler.ee = EventEmitter()
         return crawler
 
-    def connectWebDriver(self, addr: str) -> WebDriver:
+    def createUUID(self) -> str:
+        return str(uuid.uuid4())
+
+    def connectWebDriver(self, addr: str, uuid: str) -> WebDriver:
         chrome_options = webdriver.ChromeOptions()
-        prefs = {'profile.default_content_setting_values.automatic_downloads': 1}
+        prefs = {
+            'profile.default_content_setting_values.automatic_downloads': 1,
+            'download.default_directory': f"/usr/src/downloads/{uuid}"
+        }
         chrome_options.add_experimental_option("prefs", prefs)
         driver = webdriver.Remote(
             command_executor=addr,
@@ -44,9 +55,10 @@ class MarcapCrawler(Crawler[T]):
 
     async def crawling(self, dto: StockCrawlingRunCrawlingDTO) -> None:
         driver = None
-        self.ee.emit("marcapCrawler/onConnectingWebdriver")
+        uuid = self.createUUID()
+        self.ee.emit(EVENT_MARCAP_CRAWLING_ON_CONNECTING_WEBDRIVER, dto)
         try:
-            driver = self.connectWebDriver(dto.driverAddr)
+            driver = self.connectWebDriver(dto.driverAddr, uuid)
             driver.get("http://data.krx.co.kr/contents/MDC/MDI/mdiLoader/index.cmd?menuId=MDC0201020101")
             try:
                 alert = WebDriverWait(driver, timeout=3).until(EC.alert_is_present())
@@ -55,16 +67,24 @@ class MarcapCrawler(Crawler[T]):
                 print("예외발생:"+str(e))
             print("start:"+dto.startDateStr)
 
-            self.ee.emit("marcapCrawler/onStartCrawling")
+            self.ee.emit(EVENT_MARCAP_CRAWLING_ON_START_CRAWLING, dto)
             WebDriverWait(driver, timeout=20, poll_frequency=1).until(EC.element_to_be_clickable((By.CSS_SELECTOR, "#mktId_0_1")))
             date = datetime.strptime(dto.startDateStr, "%Y%m%d")
             endDate = datetime.strptime(dto.endDateStr, "%Y%m%d")
 
             while date <= endDate:
                 dateStr = date.strftime("%Y%m%d")
-                self.ee.emit("marcapCrawler/onDownloadStart")
-                downloadTask = StockCrawlingDownloadTaskDTO(dateStr, dto.market, self.ee)
-                await asyncio.create_task(self.downloadData(downloadTask, driver))
+                downloadTask = StockCrawlingDownloadTaskDTO(**{
+                    "dateStr": dateStr,
+                    "market": dto.market,
+                    "uuid": uuid,
+                    "taskId": dto.taskId,
+                    "taskUniqueId": dto.taskUniqueId
+                })
+                self.ee.emit(EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_START, downloadTask)
+                downloadObserver = DownloadObserver()
+                downloadObserver.startObserver(uuid, self.ee)
+                await self.downloadData(downloadTask, downloadObserver, driver)
                 date = date + timedelta(days=1)
         except Exception as e:
             print(f"error: {str(e)}")
@@ -72,7 +92,7 @@ class MarcapCrawler(Crawler[T]):
             if driver:
                 driver.quit()
     
-    async def downloadData(self, downloadTask: StockCrawlingDownloadTaskDTO, driver: WebDriver) -> None:
+    async def downloadData(self, downloadTask: StockCrawlingDownloadTaskDTO, downloadObserver: DownloadObserver, driver: WebDriver) -> None:
         if driver is None:
             return
         # pymitter
@@ -102,42 +122,87 @@ class MarcapCrawler(Crawler[T]):
         driver.execute_script("$(\"[data-type='csv']\").click()")
         print("wait:"+downloadTask.dateStr)
         self.isLock = True
-        
-        @self.ee.on("observer/downloadComplete")
+
+        @self.ee.on(FILE_SYSTEM_HANDLER(downloadTask.uuid))
         def downloadComplete(event: FileCreatedEvent, downloadTask: StockCrawlingDownloadTaskDTO) -> None:
             self.isLock = False
-            self.ee.emit("marcapCrawler/onDownloadComplete")
-            self.parseFile(event, downloadTask)
+            self.ee.emit(EVENT_MARCAP_CRAWLING_ON_DOWNLOAD_COMPLETE, downloadTask)
+            self.parseFile(event, downloadTask, downloadObserver)
 
         while self.isLock:
             await asyncio.sleep(1)
             print(f"isLock:{str(self.isLock)}")
+    
+    def convertFileToDto(self, path: str, dto: StockMarketCapitalResultDTO) -> None:
+        lines = []
+        with open(path, "r", encoding="utf-8") as f:
+            p = Path(f.name)
+            dto.date = p.stem
+            lines = f.readlines()
+        
+        for i in range(1, len(lines)):
+            marcap = StockMarketCapitalDTO()
+            data = lines[i].replace('"', '').split(",")
+            marcap.date = dto.date
+            marcap.market = dto.market
+            marcap.code = data[0].strip()
+            marcap.name = data[1].strip()
+            marcap.close = data[2].strip()
+            marcap.diff = data[3].strip()
+            marcap.percent = data[4].strip()
+            marcap.open = data[5].strip()
+            marcap.high = data[6].strip()
+            marcap.low = data[7].strip()
+            marcap.volume = data[8].strip()
+            marcap.price = data[9].strip()
+            marcap.marcap = data[10].strip()
+            marcap.number = data[11].strip()
+            marcap.createdAt = datetime.today()
+            marcap.updatedAt = ""
+            dto.data.append(marcap)
 
-    async def parseFile(self, event: FileCreatedEvent, downloadTask: StockCrawlingDownloadTaskDTO) -> None:
+    async def isExistFile(self, path: str, ext: str = ".csv") -> bool:
+        isExist = path.endswith(ext)
+        restTimes = 3
+        while not isExist and restTimes >= 0:
+            await asyncio.sleep(1)
+            isExist = path.endswith(ext)
+            restTimes -= 1
+        return isExist
+    
+    async def parseFile(self, event: FileCreatedEvent, downloadTask: StockCrawlingDownloadTaskDTO, downloadObserver: DownloadObserver) -> None:
         isSuccess = False
-        ret = StockMarketCapitalResultDTO()
+        retdto = StockMarketCapitalResultDTO()
         try:
             date = downloadTask.dateStr
             market = downloadTask.market
-            if not event.src_path.endswith(".csv"):
-                await asyncio.sleep(1)
-                name = os.path.basename(event.src_path).split(".")[0]
-                if not os.path.exists(os.path.dirname(event.src_path)+name+".csv"):
-                    ret.result = "fail"
-                    raise Exception('')
+            retdto.date = date
+            retdto.market = market
+            isExist = await self.isExistFile(event.src_path)
+            if not isExist:
+                raise Exception('file Not Exist')
             print("created: " + date)
-
-            dest_path = f'{os.path.dirname(event.src_path)}/{date}.csv'
             await asyncio.sleep(0.5)
-            changeCharSet(event.src_path)
+            dest_path = f'{os.path.dirname(event.src_path)}/{date}.csv'
+            self.changeCharSet(event.src_path)
             os.rename(event.src_path, dest_path)
-
-            ret.readData(dest_path, market)
-            ret.result = "success"
+            self.convertFileToDto(dest_path, retdto)
+            retdto.result = "success"
             isSuccess = True
         except Exception as e:
-            print(e)
-            ret.result = "fail"
+            retdto.result = "fail"
+            retdto.errorMsg = str(e)
         finally:
-            downloadTask.stopObserver()
-            self.ee.emit("marcapCrawler/onParsingComplete", isSuccess, ret)
+            downloadObserver.stopObserver()
+            self.ee.emit(EVENT_MARCAP_CRAWLING_ON_PARSING_COMPLETE, isSuccess, retdto, downloadTask)
+
+    def changeCharSet(self, path: str) -> None:
+        lines = None
+        with open(path, "r", encoding="euc-kr") as f:
+            lines = f.readlines()
+        with open(path, 'w', encoding="utf-8") as f:
+            f.writelines(lines)
+
+
+ee = EventEmitter()
+marcapCrawler: MarcapCrawler = MarcapCrawler(ee)
